requirements.txt
streamlit
pandas
yfinance
openai>=1.0.0
python-dotenv
plotly